PISHPS-348: fixed memory issue of ExpireOrderTaskHandler; fixed stan error attribute not repeatable
