+++ src/Resources/config/services/scheduled_tasks.xml
@@ -22,6 +22,12 @@
             <tag name="messenger.message_handler"/>
         </service>
 
+        <service id="Kiener\MolliePayments\ScheduledTask\Subscription\RenewalReminder\RenewalReminderTaskDevHandler">
+            <argument type="service" id="Kiener\MolliePayments\Repository\ScheduledTask\ScheduledTaskRepository"/>
+            <argument type="service" id="Kiener\MolliePayments\Components\Subscription\SubscriptionManager"/>
+            <argument type="service" id="mollie_payments.logger"/>
+            <tag name="messenger.message_handler"/>
+        </service>
 
         <service id="Kiener\MolliePayments\ScheduledTask\OrderStatus\ExpireOrderTask">
             <tag name="shopware.scheduled.task"/>
