+++ src/ScheduledTask/Subscription/RenewalReminder/RenewalReminderTaskDevHandler.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Kiener\MolliePayments\ScheduledTask\Subscription\RenewalReminder;
+
+#[\Symfony\Component\Messenger\Attribute\AsMessageHandler(handles: RenewalReminderTaskDev::class)]
+class RenewalReminderTaskDevHandler extends AbstractRenewalReminderTaskHandler
+{
+}
