+++ src/ScheduledTask/Subscription/RenewalReminder/RenewalReminderTaskHandler.php
@@ -1,70 +1,9 @@
-<?php declare(strict_types=1);
+<?php
+declare(strict_types=1);
 
 namespace Kiener\MolliePayments\ScheduledTask\Subscription\RenewalReminder;
 
-use Kiener\MolliePayments\Components\Subscription\SubscriptionManager;
-use Kiener\MolliePayments\Repository\ScheduledTask\ScheduledTaskRepositoryInterface;
-use Psr\Log\LoggerInterface;
-use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\MessageQueue\ScheduledTask\ScheduledTaskHandler;
-
 #[\Symfony\Component\Messenger\Attribute\AsMessageHandler(handles: RenewalReminderTask::class)]
-#[\Symfony\Component\Messenger\Attribute\AsMessageHandler(handles: RenewalReminderTaskDev::class)]
-class RenewalReminderTaskHandler extends ScheduledTaskHandler
+class RenewalReminderTaskHandler extends AbstractRenewalReminderTaskHandler
 {
-    /**
-     * @var SubscriptionManager
-     */
-    private $subscriptionManager;
-
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
-
-
-    /**
-     * @param ScheduledTaskRepositoryInterface $scheduledTaskRepository
-     * @param SubscriptionManager $subscriptionManager
-     * @param LoggerInterface $logger
-     */
-    public function __construct(ScheduledTaskRepositoryInterface $scheduledTaskRepository, SubscriptionManager $subscriptionManager, LoggerInterface $logger)
-    {
-        /** @phpstan-ignore-next-line  */
-        parent::__construct($scheduledTaskRepository->getRepository(), $logger);
-
-        $this->subscriptionManager = $subscriptionManager;
-        $this->logger = $logger;
-    }
-
-    /**
-     * @return iterable<mixed>
-     */
-    public static function getHandledMessages(): iterable
-    {
-        return [
-            RenewalReminderTask::class,
-            RenewalReminderTaskDev::class,
-        ];
-    }
-
-    /**
-     *
-     */
-    public function run(): void
-    {
-        try {
-            $this->logger->info('Starting Subscription Renewal Reminder from Scheduled Tasks.');
-
-            $context = Context::createDefaultContext();
-
-            $remindedCount = $this->subscriptionManager->remindSubscriptionRenewal($context);
-
-            $this->logger->debug($remindedCount . ' subscriptions renewal reminders have been processed successfully!');
-        } catch (\Throwable $ex) {
-            $this->logger->error(
-                'Error when running Scheduled Task for Subscription Renewal Reminders. ' . $ex->getMessage()
-            );
-        }
-    }
 }
